refactor(chat-store): extract receipt coordination
